+++ java-io-base/src/main/java/nbbrd/io/AbstractIOIterator.java
@@ -16,7 +16,7 @@
  */
 package nbbrd.io;
 
-import org.checkerframework.checker.nullness.qual.Nullable;
+import org.jspecify.annotations.Nullable;
 
 import java.io.IOException;
 import java.util.NoSuchElementException;