+++ java-io-base/src/main/java/nbbrd/io/sys/OS.java
@@ -17,7 +17,7 @@
 package nbbrd.io.sys;
 
 import lombok.NonNull;
-import org.checkerframework.checker.nullness.qual.Nullable;
+import org.jspecify.annotations.Nullable;
 
 import java.util.Locale;