+++ java-io-base/src/main/java/nbbrd/io/text/Property.java
@@ -17,7 +17,7 @@
 package nbbrd.io.text;
 
 import lombok.NonNull;
-import org.checkerframework.checker.nullness.qual.Nullable;
+import org.jspecify.annotations.Nullable;
 
 import java.util.Map;
 import java.util.Objects;