+++ java-io-http/src/main/java/nbbrd/io/http/HttpAuthScheme.java
@@ -2,5 +2,5 @@ package nbbrd.io.http;
 
 public enum HttpAuthScheme {
 
-    BASIC, NONE
+    NONE, BASIC, BEARER
 }