+++ java-io-http/src/main/java/nbbrd/io/http/HttpHeadersBuilder.java
@@ -2,7 +2,7 @@ package nbbrd.io.http;
 
 import lombok.NonNull;
 import nbbrd.design.BuilderPattern;
-import org.checkerframework.checker.nullness.qual.Nullable;
+import org.jspecify.annotations.Nullable;
 
 import java.util.*;
 import java.util.function.BiPredicate;