+++ java-io-http/src/main/java/nbbrd/io/http/HttpResponseException.java
@@ -1,7 +1,7 @@
 package nbbrd.io.http;
 
 import lombok.NonNull;
-import org.checkerframework.checker.nullness.qual.Nullable;
+import org.jspecify.annotations.Nullable;
 
 import java.io.IOException;
 import java.util.Collections;