+++ CHANGELOG.md
@@ -7,6 +7,21 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
 
+## [0.0.35] - 2025-11-07
+
+### Added
+
+- Add support for Bearer authentication scheme in HttpClient
+
+### Fixed
+
+- Fix authentication errors in HttpClient
+
+### Changed
+
+- Migrate OSSRH to Central Portal
+- Remove preemptiveAuth option in HttpContext
+
 ## [0.0.34] - 2025-05-08
 
 ### Added
@@ -303,7 +318,8 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 - Initial release
 
-[Unreleased]: https://github.com/nbbrd/java-io-util/compare/v0.0.34...HEAD
+[Unreleased]: https://github.com/nbbrd/java-io-util/compare/v0.0.35...HEAD
+[0.0.35]: https://github.com/nbbrd/java-io-util/compare/v0.0.34...v0.0.35
 [0.0.34]: https://github.com/nbbrd/java-io-util/compare/v0.0.33...v0.0.34
 [0.0.33]: https://github.com/nbbrd/java-io-util/compare/v0.0.32...v0.0.33
 [0.0.32]: https://github.com/nbbrd/java-io-util/compare/v0.0.31...v0.0.32